+++ src/ska_dataproduct_api/components/store/store_factory.py
@@ -1,11 +1,7 @@
-"""This module contains a function to select the appropriate dataproduct store class based on
-Elasticsearch availability."""
+"""This module contains a function to select the appropriate dataproduct store class."""
 import logging
 from typing import Union
 
-from ska_dataproduct_api.components.search.elasticsearch.elasticsearch import (
-    ElasticsearchMetadataStore,
-)
 from ska_dataproduct_api.components.search.in_memory.in_memory_search import (
     InMemoryDataproductSearch,
 )
@@ -14,15 +10,10 @@ from ska_dataproduct_api.components.store.in_memory.in_memory import (
 )
 from ska_dataproduct_api.components.store.persistent.postgresql import (
     PGMetadataStore,
+    PGSearchStore,
     PostgresConnector,
 )
 from ska_dataproduct_api.configuration.settings import (
-    ELASTICSEARCH_HOST,
-    ELASTICSEARCH_INDICES,
-    ELASTICSEARCH_METADATA_SCHEMA_FILE,
-    ELASTICSEARCH_PASSWORD,
-    ELASTICSEARCH_PORT,
-    ELASTICSEARCH_USER,
     POSTGRESQL_ANNOTATIONS_TABLE_NAME,
     POSTGRESQL_DBNAME,
     POSTGRESQL_HOST,
@@ -76,50 +67,21 @@ def select_metadata_store_class() -> Union[PGMetadataStore, InMemoryVolumeIndexM
 
 def select_search_store_class(
     metadata_store: Union[PGMetadataStore, InMemoryVolumeIndexMetadataStore],
-) -> Union[ElasticsearchMetadataStore, InMemoryDataproductSearch]:
+) -> Union[InMemoryDataproductSearch]:
     """
-    Selects the appropriate dataproduct search store class based on Elasticsearch availability.
+    Selects the appropriate dataproduct search store class.
 
-    This function attempts to connect to Elasticsearch. If the connection is successful,
-    an instance of `ElasticsearchMetadataStore` is returned. Otherwise, a warning message
-    is logged and an instance of `InMemoryDataproductSearch` is returned for in-memory storage.
-
-    Returns:
-        Union[ElasticsearchMetadataStore, InMemoryDataproductSearch]: An instance of either
-            `ElasticsearchMetadataStore` or `InMemoryDataproductSearch` depending on Elasticsearch
-            availability.
     """
-    elastic_store_instance = ElasticsearchMetadataStore(
-        host=ELASTICSEARCH_HOST,
-        port=ELASTICSEARCH_PORT,
-        user=ELASTICSEARCH_USER,
-        password=ELASTICSEARCH_PASSWORD,
-        indices=ELASTICSEARCH_INDICES,
-        schema=ELASTICSEARCH_METADATA_SCHEMA_FILE,
-        metadata_store=metadata_store,
-    )
-    elastic_store_instance.check_and_reconnect()
 
     try:
-        elastic_store_instance = ElasticsearchMetadataStore(
-            host=ELASTICSEARCH_HOST,
-            port=ELASTICSEARCH_PORT,
-            user=ELASTICSEARCH_USER,
-            password=ELASTICSEARCH_PASSWORD,
-            indices=ELASTICSEARCH_INDICES,
-            schema=ELASTICSEARCH_METADATA_SCHEMA_FILE,
-            metadata_store=metadata_store,
+        pg_search_store = PGSearchStore(
+            db=metadata_store.db,
+            science_metadata_table_name=metadata_store.science_metadata_table_name,
+            annotations_table_name=metadata_store.annotations_table_name,
         )
-        if elastic_store_instance.host and elastic_store_instance.check_and_reconnect():
-            logger.info("Elasticsearch reachable, setting search store to ElasticSearch")
-            return elastic_store_instance
+        logger.info("PGSearchStore reachable, setting search store to PGSearchStore")
+        return pg_search_store
     except Exception as exception:  # pylint: disable=broad-exception-caught
-        logger.error("Failed to connect to Elasticsearch with exception: %s", exception)
+        logger.error("Failed to connect to PGSearchStore with exception: %s", exception)
         logger.warning("Using InMemoryDataproductSearch search.")
         return InMemoryDataproductSearch(metadata_store)
-
-    logger.warning(
-        "Elasticsearch not available, setting search store to InMemoryDataproductSearch\
-                    store."
-    )
-    return InMemoryDataproductSearch(metadata_store)